+++ Samples~/Scripts/SaintsEditor/FieldTagSub.cs
@@ -1,0 +1,18 @@
+using System;
+
+namespace SaintsField.Samples.Scripts.SaintsEditor
+{
+    public class FieldTagSub : SaintsMonoBehaviour
+    {
+        [Serializable]
+        public struct MyStruct
+        {
+            public string myName;
+        }
+
+        [LabelText("<field.myName />")]
+        [AboveText("<field.myName />")]
+        [BelowText("<field.myName />")]
+        public MyStruct myStruct;
+    }
+}